refactor(formatter): move `is_supported_source_type` to `oxc_formatter` crate (#13702)
- Renamed `is_supported_source_type` to `get_supported_source_type`
- Moved the function to `oxc_formatter`, because the language server wants to use the function too
