+++ crates/oxc_formatter/src/service/mod.rs
@@ -1,0 +1,1 @@
+pub mod source_type;
